requirements.txt
Flask
openai
pandas
numpy
azure-storage-file-datalake
pyarrow
azure-identity
azure-keyvault-secrets